Parameter Make.1-C

type t

val compare : t -> t -> int

Compare two clauses

val to_lits : t -> Index_intf.lits

Iterate on literals of the clause

val labels : t -> Index_intf.labels

Some integer labels. We assume that if c to subsume d, then labels c is a subset of labels d