Module Hashcons.MakeNonWeak

Version that uses a regular Hashtbl, rather than a weak table. Never frees memory, but might be faster.

Parameters

X : HashedType

Signature

type elt = X.t

Hashconsed objects

val hashcons : elt -> elt

Hashcons the elements

val mem : elt -> bool

Is the element present in this table?

val fresh_unique_id : unit -> int

Unique ID that will never occur again in this table (modulo 2^63...)

val stats : unit -> int * int * int * int * int * int