Module Logtk.Input_format

Input Format

type t =

| I_tptp

| I_zf

| I_tip

| I_dk

val tptp : t

val tip : t

val zf : t

val dk : t

val default : t

val pp : CCFormat.t -> t -> unit

val on_undef_id : t -> [> `Fail | `Guess ]

What to do when we have an undefined ID in the corresponding format?

val on_shadow : t -> [> `Ignore | `Warn ]

What to do when we have a shadowing decl?

val on_var : t -> [> `Default | `Infer ]

what to do when we have a variable without a type declaration?

val implicit_ty_args : t -> bool

Do we add implicit type parameters when '@' is not present?