Module Type.TPTP

include Interfaces.PRINT_DE_BRUIJN with type term := t and type t := t

type t

type term

type print_hook = int -> term CCFormat.printer -> Format.formatter -> term -> bool

User-provided hook that can be used to print terms (for composite cases) before the default printing occurs. The int argument is the De Bruijn depth in the term. A hook takes as arguments the depth and the recursive printing function that it can use to print subterms. A hook should return true if it fired, false to fall back on the default printing.

val pp_depth : ?⁠hooks:print_hook list -> int -> t CCFormat.printer

include Interfaces.PRINT with type t := t

type t

val pp : t CCFormat.printer

val to_string : t -> string

val pp_typed_var : t HVar.t CCFormat.printer

val pp_ho : CCFormat.t -> t -> unit

Basic types

val i : t

individuals

val o : t

propositions

val int : t

integers

val rat : t

rationals

val real : t

reals