Module Logtk.UnionFind

Imperative Union-Find structure

module type PAIR = sig ... end

We need to be able to hash and compare keys, and values need to form a monoid

module Make : functor (P : PAIR) -> sig ... end

Build a union-find module from a key/value specification