Module Logtk__Congruence_intf

The congruence stores a finite representation of a set of (ground) equations an inequalities. It is backtrackable, ie one can go back to a previous point as long as it is compatible with a LIFO usage.

module type S = sig ... end