Module Logtk__FV_tree.Make

Parameters

C : Logtk.Index_intf.CLAUSE

Signature

module Feature_fun : sig ... end

type feature_funs = Feature_fun.t Logtk.IArray.t

val compute_fv_c : feature_funs -> C.t -> feature_vector

Index

include Logtk.Index.SUBSUMPTION_IDX with module C = C

type t

module C = C

val name : string

val empty : unit -> t

Empty index

val add : t -> C.t -> t

Index the clause

val add_seq : t -> C.t Iter.t -> t

val add_list : t -> C.t list -> t

val remove : t -> C.t -> t

Un-index the clause

val remove_seq : t -> C.t Iter.t -> t

val retrieve_subsuming : t -> Logtk.Index_intf.lits -> Logtk.Index_intf.labels -> C.t Iter.t

Fold on a set of indexed candidate clauses, that may subsume the given clause.

val retrieve_subsuming_c : t -> C.t -> C.t Iter.t

val retrieve_subsumed : t -> Logtk.Index_intf.lits -> Logtk.Index_intf.labels -> C.t Iter.t

Fold on a set of indexed candidate clauses, that may be subsumed by the given clause

val retrieve_subsumed_c : t -> C.t -> C.t Iter.t

val retrieve_alpha_equiv : t -> Logtk.Index_intf.lits -> Logtk.Index_intf.labels -> C.t Iter.t

Retrieve clauses that are potentially alpha-equivalent to the given clause

val retrieve_alpha_equiv_c : t -> C.t -> C.t Iter.t

Retrieve clauses that are potentially alpha-equivalent to the given clause

val iter : t -> C.t Iter.t

val fold : ('a -> C.t -> 'a) -> 'a -> t -> 'a

val empty_with : feature_funs -> t

val default_feature_funs : feature_funs

val feature_funs : t -> feature_funs