Module Logtk__Int_lit.Focus

type t =

| Left of op * Z.t Logtk.Monome.Focus.t * Z.t Logtk.Monome.t

| Right of op * Z.t Logtk.Monome.t * Z.t Logtk.Monome.Focus.t

| Div of Z.t Logtk.Monome.Focus.t divides

focus on a term in one of the two monomes

val mk_left : op -> Z.t Logtk.Monome.Focus.t -> Z.t Logtk.Monome.t -> t

val mk_right : op -> Z.t Logtk.Monome.t -> Z.t Logtk.Monome.Focus.t -> t

val mk_div : ?⁠sign:bool -> Z.t -> power:int -> Z.t Logtk.Monome.Focus.t -> t

val get : lit -> Logtk.Position.t -> t option

val get_exn : lit -> Logtk.Position.t -> t

raises Invalid_argument

if the position doesn't fit

val focus_term : lit -> term -> t option

Attempt to focus on the given atomic term, if it occurs directly under the arith literal

val focus_term_exn : lit -> term -> t

val replace : t -> Z.t Logtk.Monome.t -> lit

replace a m replaces mf.coeff × mf.term with m where mf is the focused monome in a, and return the resulting literal

val term : t -> term

Focused term

val focused_monome : t -> Z.t Logtk.Monome.Focus.t

The focused monome

val opposite_monome : t -> Z.t Logtk.Monome.t option

The opposite monome (in Left and Right), if any.

val opposite_monome_exn : t -> Z.t Logtk.Monome.t

Unsafe version of opposite_monome.

raises Invalid_argument

if the literal is a Div

val is_max : ord:Logtk.Ordering.t -> t -> bool

Is the focused term maximal in the literal?

val is_strictly_max : ord:Logtk.Ordering.t -> t -> bool

Is the focused term maximal in the literal, ie is it greater than all the other terms?

val map_lit : f_m:(Z.t Logtk.Monome.t -> Z.t Logtk.Monome.t) -> f_mf:(Z.t Logtk.Monome.Focus.t -> Z.t Logtk.Monome.Focus.t) -> t -> t

Apply a function to the monomes and focused monomes

val product : t -> Z.t -> t

Product by a constant

val scale : t -> t -> t * t

Multiply the two literals by some constant so that their focused literals have the same coefficient

val scale_power : t -> int -> t

For a "divides" literal, bring it to the given power.

raises Invalid_argument

if the current power is strictly greater than the argument (cannot scale down), or if the literal is not a Div

val apply_subst : Logtk.Subst.Renaming.t -> Logtk.Subst.t -> t Logtk.Scoped.t -> t

Apply a substitution

val unify : ?⁠subst:Logtk.Unif_subst.t -> t Logtk.Scoped.t -> t Logtk.Scoped.t -> (t * t * Logtk.Unif_subst.t) Iter.t

Unify the two focused terms, and possibly other terms of their respective focused monomes; yield the new literals accounting for the unification along with the unifier

val fold_terms : ?⁠pos:Logtk.Position.t -> lit -> (t * Logtk.Position.t) Iter.t

Fold on focused terms in the literal, one by one, with the position of the focused term

val op : t -> [ `Binary of op | `Divides ]

val unfocus : t -> lit

Conversion back to a literal

val pp : t CCFormat.printer

val to_string : t -> string