Module Logtk__JP_unif

module T = Logtk.Term

module US = Logtk.Unif_subst

type subst = US.t

module S : sig ... end

val project_onesided : scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> subst OSeq.t

val imitate_onesided : scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> T.t -> subst OSeq.t

val imitate : scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> T.t -> (T.var * int) list -> subst OSeq.t

val identify : scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> T.t -> (T.var * int) list -> subst OSeq.t

val eliminate : scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> T.t -> (Logtk.Type.t Logtk.HVar.t * int) list -> subst OSeq.t

val iterate : ?⁠flex_same:bool -> scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> T.t -> (T.var * 'a) CCList.t -> Logtk.Unif_subst.t option OSeq.t

val find_disagreement : T.t -> T.t -> ((T.t * T.t) * (T.var * int) CCList.t) option

Find disagreeing subterms. This function also returns a list of variables occurring above the disagreement pair, along with the index of the argument that the disagreement pair occurs in.

val unify : scope:Logtk.Scoped.scope -> counter:int Stdlib.ref -> T.t -> T.t -> subst option OSeq.t

val unify_scoped_nonterminating : T.t Logtk.Scoped.t -> T.t Logtk.Scoped.t -> subst OSeq.t

val unify_scoped : T.t Logtk.Scoped.t -> T.t Logtk.Scoped.t -> subst option OSeq.t