Module Logtk__Literals.Pos

val at : t -> Logtk.Position.t -> term

Return the subterm at the given position, or

raises Invalid_argument

if the position is not valid

val lit_at : t -> Logtk.Position.t -> Logtk.Literal.t * Logtk.Position.t

Lookup which literal the position is about, return it and the rest of the position.

raises Invalid_argument

if the position is not valid

val replace : t -> at:Logtk.Position.t -> by:term -> unit

In-place modification of the array, in which the subterm at given position is replaced by the by term.

raises Invalid_argument

if the position is not valid

val idx : Logtk.Position.t -> int

Index in the literal array

raises Invalid_argument

if the position is incorrect

val tail : Logtk.Position.t -> Logtk.Position.t

sub-position

raises Invalid_argument

if the position is incorrect

val cut : Logtk.Position.t -> int * Logtk.Position.t

Index + literal position