Module Logtk__Proof.Rule

A rule is a name for some specific inference or transformation rule that is used to deduce formulas from other formulas.

type t = rule

val pp : t CCFormat.printer

val name : t -> string

val mk : string -> t

val mkf : ('a, Stdlib.Format.formatter, unit, t) Stdlib.format4 -> 'a