Module Term.Rule

type t = rule

val lhs : t -> term

val rhs : t -> term

val vars : t -> Logtk.Term.VarSet.t

val vars_l : t -> Logtk.Type.t Logtk.HVar.t list

val head_id : t -> Logtk.ID.t

val args : t -> term list

val arity : t -> int

val proof : t -> proof

val as_lit : t -> Logtk.Literal.t

val make_const : proof:Logtk.Proof.t -> Logtk.ID.t -> Logtk.Type.t -> term -> t

make_const id ty rhs is the same as T.const id ty --> rhs

val make : proof:Logtk.Proof.t -> Logtk.ID.t -> Logtk.Type.t -> term list -> term -> t

make id ty args rhs is the same as T.app (T.const id ty) args --> rhs

include Logtk.Interfaces.HASH with type t := t

include Logtk.Interfaces.EQ

type t

val equal : t -> t -> bool

val hash : t -> int

include Logtk.Interfaces.ORD with type t := t

type t

val compare : t -> t -> int

include Logtk.Interfaces.PRINT with type t := t

type t

val pp : t CCFormat.printer

val to_string : t -> string