Module Logtk__Unif

Unification and Matching

type unif_subst = Logtk.Unif_subst.t

type subst = Logtk.Subst.t

type term = Logtk.InnerTerm.t

type ty = Logtk.InnerTerm.t

type 'a sequence = ('a -> unit) -> unit

exception Fail

Raised when a unification/matching attempt fails

val _allow_pattern_unif : bool Stdlib.ref

val _unif_bool : bool Stdlib.ref

val norm_logical_disagreements : Logtk.Builtin.t -> Logtk.Term.t list -> Logtk.Term.t list -> Logtk.Term.t list * Logtk.Term.t list

val occurs_check : depth:int -> subst -> Logtk.InnerTerm.t Logtk.HVar.t Logtk.Scoped.t -> Logtk.InnerTerm.t Logtk.Scoped.t -> bool

val unif_array_com : ?⁠size:[ `Same | `Smaller ] -> 'subst -> op:('subst -> 'a Logtk.Scoped.t -> 'a Logtk.Scoped.t -> 'subst Iter.t) -> 'a array Logtk.Scoped.t -> 'a array Logtk.Scoped.t -> 'subst Iter.t

Generic unification over two arrays (of the same size, or the first one must be smaller or equal)

val unif_list : 'subst -> op:('subst -> 'a Logtk.Scoped.t -> 'a Logtk.Scoped.t -> 'subst Iter.t) -> 'a list Logtk.Scoped.t -> 'a list Logtk.Scoped.t -> 'subst Iter.t

Generic unification over two lists (of the same size)

val unif_list_com : ?⁠size:[ `Same | `Smaller ] -> 'subst -> op:('subst -> 'a Logtk.Scoped.t -> 'a Logtk.Scoped.t -> 'subst Iter.t) -> 'a list Logtk.Scoped.t -> 'a list Logtk.Scoped.t -> 'subst Iter.t

Generic unification over two lists (of the same size or smaller)

val pair_lists_right : term -> term list -> term -> term list -> term list * term list

in HO, we have f1 l1 and f2 l2, where application is left-associative. we need to unify from the right (the outermost application is on the right) so this returns pairs to unify (including heads).

val pair_lists_left : term list -> term -> term list -> term -> term list * term list

in HO, we have l1 -> ret1 and l2 -> ret2, where -> is right-associative. we need to unify from the left, so this returns pairs to unify (including return types).

Signatures

module type S = Logtk.Unif_intf.S

Base (scoped terms)

module Inner : S with type term = Logtk.InnerTerm.t and type ty = Logtk.InnerTerm.t

To be used only on terms without InnerTerm.Multiset constructor

Specializations

module Ty : sig ... end

module FO : sig ... end