Module Logtk__Util.Section

Debug section

type t

val full_name : t -> string

Full path to the section

val set_debug : t -> int -> unit

Debug level for section (and its descendants)

val clear_debug : t -> unit

Clear debug level (will be same as parent's)

val get_debug : t -> int option

Specific level of this section, if any

val cur_level : t -> int

Current debug level, with parent inheritance

val iter : (string * t) Iter.t

all registered sections

val root : t

Default section, with no parent

val make : ?⁠parent:t -> ?⁠inheriting:t list -> string -> t

make ?parent ?inheriting name makes a new section with the given name. It has a parent (default root), used to give it a name. It can also have a list of sections it inherits from. Unless specified explicitly otherwise (using set_debug, the level of the section will be the max level of its parent and its inherited sections.