Module Logtk_parsers.Util_zf

Utils for ZF

type parse_cache

Cache that remembers the set of files that have been parsed so far

val create_parse_cache : unit -> parse_cache

type parser_res = (Logtk.UntypedAST.statement Iter.t, string) CCResult.t

type 'a parser_ = 'a -> parser_res

val parse_lexbuf : ?⁠cache:parse_cache -> ?⁠recursive:bool -> Stdlib.Lexing.lexbuf parser_

parse lexbuf.

parameter recursive

if true, parse includes too, relative to current dir. Default true.

parameter cache

parse cache to use if recursive = true

val parse_stdin : unit parser_

parse stdin. Does not expand includes.

val parse_file : ?⁠cache:parse_cache -> ?⁠recursive:bool -> string parser_

parse_file ~recursive file parses file

parameter recursive

if true, recursively parse includes. Default true.

parameter cache

parse cache to use if recursive = true