Module Logtk_parsers__Parsing_utils

Utils for Parsing in various Formats (High-Level Interface)

type 'a or_error = ('a, string) CCResult.t

val parse_tptp : string -> Logtk.UntypedAST.statement Iter.t or_error

Parse the given file as TPTP

val parse_tip : string -> Logtk.UntypedAST.statement Iter.t or_error

Parse the given file as TIP

val guess_input : string -> Logtk.Input_format.t

Guess input from file extension

val input_of_file : string -> Logtk.Input_format.t

Choose input for this file based on Options.input and guess_input. This is the recommended way of picking the input format for a given file.

val parse_file : Logtk.Input_format.t -> string -> Logtk.UntypedAST.statement Iter.t or_error

parse_file fmt file parses the file using the parser for the given input format.