Module Literal.View

Specific views

val as_eqn : t -> (term * term * bool) option

val get_eqn : t -> Position.t -> (term * term * bool) option

View of a Prop or Equation literal, oriented by the position. If the position selects its left term, return l, r, otherwise r, l. for propositions it will always be p, true.

returns

None for other literals

raises Invalid_argument

if the position doesn't match the literal.

val get_arith : t -> Int_lit.t option

Extract an arithmetic literal

val focus_arith : t -> Position.t -> Int_lit.Focus.t option

Focus on a specific term in an arithmetic literal. The focused term is removed from its monome, and its coefficient is returned.

val unfocus_arith : Int_lit.Focus.t -> t

val get_rat : t -> Rat_lit.t option

Extract an arithmetic literal

val focus_rat : t -> Position.t -> Rat_lit.Focus.t option

Focus on a specific term in an arithmetic literal. The focused term is removed from its monome, and its coefficient is returned.

val unfocus_rat : Rat_lit.Focus.t -> t